# Optimize Poseidon from *big.Int to goff