--- a/poseidon/poseidon_test.go
+++ b/poseidon/poseidon_test.go
@@ -5,6 +5,7 @@ import (
 	"math/big"
 	"testing"
 
+	"github.com/iden3/go-iden3-crypto/ff"
 	"github.com/iden3/go-iden3-crypto/utils"
 	"github.com/stretchr/testify/assert"
 	"golang.org/x/crypto/blake2b"
@@ -16,46 +17,46 @@ func TestBlake2bVersion(t *testing.T) {
 }
 
 func TestPoseidon(t *testing.T) {
-	b1 := big.NewInt(int64(1))
-	b2 := big.NewInt(int64(2))
-	h, err := Hash([]*big.Int{b1, b2})
+	b1 := utils.NewElement().SetUint64(1)
+	b2 := utils.NewElement().SetUint64(2)
+	h, err := Hash([]*ff.Element{b1, b2})
 	assert.Nil(t, err)
 	assert.Equal(t, "4932297968297298434239270129193057052722409868268166443802652458940273154855", h.String())
 
-	b3 := big.NewInt(int64(3))
-	b4 := big.NewInt(int64(4))
-	h, err = Hash([]*big.Int{b3, b4})
+	b3 := utils.NewElement().SetUint64(3)
+	b4 := utils.NewElement().SetUint64(4)
+	h, err = Hash([]*ff.Element{b3, b4})
 	assert.Nil(t, err)
 	assert.Equal(t, "4635491972858758537477743930622086396911540895966845494943021655521913507504", h.String())
 
 	msg := []byte("Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.")
 	n := 31
-	msgElems := make([]*big.Int, 0, len(msg)/n+1)
+	msgElems := make([]*ff.Element, 0, len(msg)/n+1)
 	for i := 0; i < len(msg)/n; i++ {
 		v := new(big.Int)
 		utils.SetBigIntFromLEBytes(v, msg[n*i:n*(i+1)])
-		msgElems = append(msgElems, v)
+		msgElems = append(msgElems, utils.NewElement().SetBigInt(v))
 	}
 	if len(msg)%n != 0 {
 		v := new(big.Int)
 		utils.SetBigIntFromLEBytes(v, msg[(len(msg)/n)*n:])
-		msgElems = append(msgElems, v)
+		msgElems = append(msgElems, utils.NewElement().SetBigInt(v))
 	}
 	hmsg, err := Hash(msgElems)
 	assert.Nil(t, err)
 	assert.Equal(t, "16019700159595764790637132363672701294192939959594423814006267756172551741065", hmsg.String())
 
 	msg2 := []byte("Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum. Lorem ipsum dolor sit amet.")
-	msg2Elems := make([]*big.Int, 0, len(msg2)/n+1)
+	msg2Elems := make([]*ff.Element, 0, len(msg2)/n+1)
 	for i := 0; i < len(msg2)/n; i++ {
 		v := new(big.Int)
 		utils.SetBigIntFromLEBytes(v, msg2[n*i:n*(i+1)])
-		msg2Elems = append(msg2Elems, v)
+		msg2Elems = append(msg2Elems, utils.NewElement().SetBigInt(v))
 	}
 	if len(msg2)%n != 0 {
 		v := new(big.Int)
 		utils.SetBigIntFromLEBytes(v, msg2[(len(msg2)/n)*n:])
-		msg2Elems = append(msg2Elems, v)
+		msg2Elems = append(msg2Elems, utils.NewElement().SetBigInt(v))
 	}
 	hmsg2, err := Hash(msg2Elems)
 	assert.Nil(t, err)
@@ -67,29 +68,41 @@ func TestPoseidon(t *testing.T) {
 }
 
 func TestPoseidonBrokenChunks(t *testing.T) {
-	h1, err := Hash([]*big.Int{big.NewInt(0), big.NewInt(1), big.NewInt(2), big.NewInt(3), big.NewInt(4),
-		big.NewInt(5), big.NewInt(6), big.NewInt(7), big.NewInt(8), big.NewInt(9)})
+	h1, err := Hash([]*ff.Element{utils.NewElement().SetUint64(0), utils.NewElement().SetUint64(1), utils.NewElement().SetUint64(2), utils.NewElement().SetUint64(3), utils.NewElement().SetUint64(4),
+		utils.NewElement().SetUint64(5), utils.NewElement().SetUint64(6), utils.NewElement().SetUint64(7), utils.NewElement().SetUint64(8), utils.NewElement().SetUint64(9)})
 	assert.Nil(t, err)
-	h2, err := Hash([]*big.Int{big.NewInt(5), big.NewInt(6), big.NewInt(7), big.NewInt(8), big.NewInt(9),
-		big.NewInt(0), big.NewInt(1), big.NewInt(2), big.NewInt(3), big.NewInt(4)})
+	h2, err := Hash([]*ff.Element{utils.NewElement().SetUint64(5), utils.NewElement().SetUint64(6), utils.NewElement().SetUint64(7), utils.NewElement().SetUint64(8), utils.NewElement().SetUint64(9),
+		utils.NewElement().SetUint64(0), utils.NewElement().SetUint64(1), utils.NewElement().SetUint64(2), utils.NewElement().SetUint64(3), utils.NewElement().SetUint64(4)})
 	assert.Nil(t, err)
 	assert.NotEqual(t, h1, h2)
 }
 
 func TestPoseidonBrokenPadding(t *testing.T) {
-	h1, err := Hash([]*big.Int{big.NewInt(1)})
+	h1, err := Hash([]*ff.Element{utils.NewElement().SetUint64(1)})
 	assert.Nil(t, err)
-	h2, err := Hash([]*big.Int{big.NewInt(1), big.NewInt(0)})
+	h2, err := Hash([]*ff.Element{utils.NewElement().SetUint64(1), utils.NewElement().SetUint64(0)})
 	assert.Nil(t, err)
 	assert.NotEqual(t, h1, h2)
 }
 
 func BenchmarkPoseidon(b *testing.B) {
-	b12 := big.NewInt(int64(12))
-	b45 := big.NewInt(int64(45))
-	b78 := big.NewInt(int64(78))
-	b41 := big.NewInt(int64(41))
-	bigArray4 := []*big.Int{b12, b45, b78, b41}
+	b12 := utils.NewElement().SetUint64(12)
+	b45 := utils.NewElement().SetUint64(45)
+	b78 := utils.NewElement().SetUint64(78)
+	b41 := utils.NewElement().SetUint64(41)
+	bigArray4 := []*ff.Element{b12, b45, b78, b41}
+
+	for i := 0; i < b.N; i++ {
+		Hash(bigArray4)
+	}
+}
+
+func BenchmarkPoseidonLarge(b *testing.B) {
+	b12 := utils.NewElement().SetString("11384336176656855268977457483345535180380036354188103142384839473266348197733")
+	b45 := utils.NewElement().SetString("11384336176656855268977457483345535180380036354188103142384839473266348197733")
+	b78 := utils.NewElement().SetString("11384336176656855268977457483345535180380036354188103142384839473266348197733")
+	b41 := utils.NewElement().SetString("11384336176656855268977457483345535180380036354188103142384839473266348197733")
+	bigArray4 := []*ff.Element{b12, b45, b78, b41}
 
 	for i := 0; i < b.N; i++ {
 		Hash(bigArray4)